Restructure repository
This is an attempt to make the files in this repo easier to work with.

There is no separate "client" directory, as we don't yet have a client
which is distinct form the server. This should be revisited in future.

Models, controllers, and views have been not been grouped into domain
areas as currently the domain is so tiny that everything relates to the
single "job application" domain. This should be revisited in future.

It is hoped that this change will facilitate a general cleanup of the
code, including things like extracting controllers from routes,
introducting a top-level router, breaking up the app.js file, etc.